UserRouter -> userRouter 소문자로 변경, 등록된 수정, 제품 상세 페이지에서 신청(트랜잭션 저장) 기능 추가, 라우터에 지정하지 않은 주소 이동 시 NotFound 페이지 추가, 제품 상세 페이지, 제품 등록 페이지, 트랜잭션 리스트 조회 추가 회원가입 후 로그인 화면으로 이동 추가,
