+++ backend/src/main/java/com/example/oauthjwt/dto/request/TransactionUpdateRequest.java
@@ -1,11 +1,10 @@
 package com.example.oauthjwt.dto.request;
 
-import com.example.oauthjwt.entity.ServiceProduct;
+import com.example.oauthjwt.entity.Product;
 import com.example.oauthjwt.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
 
 @Data
 @NoArgsConstructor
@@ -13,5 +12,5 @@ import lombok.experimental.SuperBuilder;
 public class TransactionUpdateRequest extends TransactionRequest {
     private Long id; // 트랜잭션 id
     private User user;
-    private ServiceProduct serviceProduct;
+    private Product product;
 }
