+++ backend/src/main/java/com/example/oauthjwt/service/impl/TransactionServiceImpl.java
@@ -1,21 +1,19 @@
 package com.example.oauthjwt.service.impl;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.example.oauthjwt.dto.request.TransactionUpdateRequest;
 import com.example.oauthjwt.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.example.oauthjwt.dto.request.TransactionRequest;
 import com.example.oauthjwt.dto.response.TransactionResponse;
-import com.example.oauthjwt.entity.ServiceProduct;
+import com.example.oauthjwt.entity.Product;
 import com.example.oauthjwt.entity.TransactionStatus;
 import com.example.oauthjwt.entity.Transaction;
 import com.example.oauthjwt.entity.User;
-import com.example.oauthjwt.repository.ServiceProductRepository;
+import com.example.oauthjwt.repository.ProductRepository;
 import com.example.oauthjwt.repository.TransactionRepository;
 import com.example.oauthjwt.repository.UserRepository;
 import com.example.oauthjwt.service.TransactionService;
@@ -27,7 +25,7 @@ import org.springframework.web.server.ResponseStatusException;
 @RequiredArgsConstructor
 public class TransactionServiceImpl implements TransactionService {
   private final UserRepository userRepository;
-  private final ServiceProductRepository serviceProductRepository;
+  private final ProductRepository productRepository;
   private final TransactionRepository transactionRepository;
 
   private final UserService userService;
@@ -37,14 +35,17 @@ public class TransactionServiceImpl implements TransactionService {
     // 검증
     User user = userRepository.findById(transactionRequest.getUserId())
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "유저 정보가 존재하지 않습니다."));
-    ServiceProduct serviceProduct = serviceProductRepository.findById(transactionRequest.getProductId())
+    Product product = productRepository.findById(transactionRequest.getProductId())
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "제품 정보가 존재하지 않습니다."));
-    TransactionStatus transactionStatus = TransactionStatus.parseTransactionType(transactionRequest.getTransactionStatus().toUpperCase());
+    TransactionStatus transactionStatus = TransactionStatus.parseTransactionType(transactionRequest.getStatus().toUpperCase());
+    if(user.getId().equals(product.getOwner().getId())){
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "자신이 올린 글에는 요청을 할 수없습니다.");
+    }
     if(transactionStatus == null){
       throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "허용되지 않는 타입입니다.");
     }
     // 엔티티 생성
-    Transaction transaction = Transaction.of(transactionRequest, user, serviceProduct, transactionStatus);
+    Transaction transaction = Transaction.of(transactionRequest, user, product, transactionStatus);
     // 저장 및 반환
     Transaction result = transactionRepository.save(transaction);
     return TransactionResponse.toDto(result);
@@ -69,12 +70,12 @@ public class TransactionServiceImpl implements TransactionService {
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "트랜잭션 정보가 존재하지 않습니다."));
     User user = userRepository.findById(transactionRequest.getUserId())
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "유저 정보가 존재하지 않습니다."));
-    ServiceProduct serviceProduct = serviceProductRepository.findById(transactionRequest.getProductId())
+    Product product = productRepository.findById(transactionRequest.getProductId())
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "제품 정보가 존재하지 않습니다."));
     // 값 업데이트
-    TransactionStatus transactionStatus = TransactionStatus.parseTransactionType(transactionRequest.getTransactionStatus().toUpperCase());
+    TransactionStatus transactionStatus = TransactionStatus.parseTransactionType(transactionRequest.getStatus().toUpperCase());
     // 저장 및 반환
-    Transaction result = transactionRepository.save(transaction.setUpdateValue(transactionRequest, user, serviceProduct, transactionStatus));
+    Transaction result = transactionRepository.save(transaction.setUpdateValue(transactionRequest, user, product, transactionStatus));
     return TransactionResponse.toDto(result);
   }
 }
