+++ client-app/src/router/productRouter.jsx
@@ -1,0 +1,29 @@
+import React, {lazy} from "react"
+
+const Read = lazy(() => import("../component/product/Read.jsx"))
+const List = lazy(() => import("../component/product/List.jsx"))
+const Save = lazy(() => import("../component/product/Save.jsx"))
+const Modify = lazy(() => import("../component/product/Modify.jsx"))
+
+
+const productRouter = () =>{
+    return [
+    {
+        path: "read/:id",
+        element: <Read/>
+    },
+    {
+        path: "list",
+        element: <List/>
+    },
+    {
+        path: "save",
+        element: <Save/>
+    },
+    {
+        path: "modify/:id",
+        element: <Modify/>
+    },
+]}
+
+export default productRouter
